Flexible Analytics Manager
Allow analytics provider plugins to specify template tag names to allow
templates to specify where on the DOM individual providers are
rendered.  (for example: google tag manager requires a data layer in
the head however, the core tag manager snippet might be best placed in
the body of the document).