Making ArticleServiceV2 and ArticleRenderer smarter about rendering "read more" links.
ArticleServiceV2 now sets a readMore flag on the article context. This flag can be passed to the get(), getSingle() and getAll() functions. The readMore flag defaults to false for get() and getSingle(), and true for getAll(). ArticleRenderer checks for this flag on the ArticleContext to determine if the article body content should be truncated, and "read more" links rendered.