Fixes issue with Localization not working correctly in 0.5.0
Localization service expected the locale to be in the form language_countrycode, however when loading from the files they come iin as language-country code.  When the localization service searched through its record of supported languages, it could not find en-us due to the fact that its storage object had the id en_us.

The fix adds an if-statement that converts en-us to en_us (works for any language/country code combination)