The database now supports deletions.  At the start of save, it takes the object from the database and the payload from the post.  It will then iterate through each key in the DB object.  If it finds that the key has the language/country values, it will delete the country as a whole.
(we can expand this to respect plugin if we want to.)

It will recursively (sorta) delete the empty parent structure, so if your key only has en-US translations, it will delete first the US object, then checks if the en is empty, which it would be, then that Object is deleted, and then the key is checked, if it only has one property (isKey) it will be deleted as well.

This last bit was to provide for cleanup to make sure all remains of deleted localizations are removed from the object structure.

The normal flow is then respected, at which point every object that was submitted to the backend will be created and shoved onto the DB object and then written back to the DB.