Adding Transaction Utilities
Adding the ability to run transaction blocks against things in the session.  This will basically create a mutex around a block of code.  Given a function (whose parameters are already bound) it will call this with a callback (in the standard or async versions) and then call it without a callback in the sync version.