Added a few quality of life tweaks to the middleware.  Added a todo to fix the system setup check.  Fixed form controller so it doesnt explode.
Fixed redirects so they work better with KOA.  Fixed a bug in access management that now prevents an infinite redirect loop.

Cleaned up some login action code and finally cleared out the old PB Blog posts and made way for the ability to change that via localization keys so that we can update that message on the fly.