Updating Router to look more like the original 0.8.3 change set.  Removed the old handler function from 0.8.1
In serve public content, we correctly attach the data to controllerResult.

In System for handling of public routes and node modules (Which is deprecated for all but core PB routes/pages) there was an issue where a callback function was not awaiting compleition before it was allow to continue the middleware stack.  This was the reason for the error that we saw in the logs - all be it in a very long, round about, and obtuse way.

It would not set the data on the req.controllerResult.  When in Debug mode, that variable is referenced in recordEndTime middleware.  That would error and cause it to fallback into the error handler which would render a new page and call through to the error controller.  Which would go through and reset headers.  By this point, that function that didnt get awaited correctly would have set the headers and sent the request - fulfilling the front end's needs.

In Summary: Correctly sets the controller result int the serving of public routes.  Correctly awaits the completion of serving public content before continuing the middleware stack.  Remove dead code.